@@ -8,10 +8,15 @@ from farconf import parse_cli, to_dict
 from names_generator import generate_name
 from stable_baselines3.common.type_aliases import check_cast
 
-# Make sure the command configs are registered
-import learned_planner.cmd  # type: ignore
 import learned_planner.evaluate  # noqa: F401
 from learned_planner.configs.command_config import WandbCommandConfig
+from learned_planner.interp.collect_dataset import DatasetStore  # noqa: F401
+
+# Make sure the command configs are registered
+try:
+    import learned_planner.cmd  # noqa: F401
+except ImportError:
+    pass
 
 
 def setup_run(cfg: WandbCommandConfig) -> Path:
@@ -57,11 +62,14 @@ def setup_run(cfg: WandbCommandConfig) -> Path:
     #
     # We make a new directory with a different name ("local-files") and save our files there, so they don't get synced
     # to wandb.
-    wandb_run_dir = Path(wandb.run.dir)
-    assert wandb_run_dir.name == "files"
 
-    files_dir = wandb_run_dir.parent / "local-files"
-    files_dir.mkdir()
+    wandb_run_dir = Path(wandb.run.dir)
+    if wandb_kwargs["mode"] == "disabled":
+        files_dir = wandb_run_dir / "local-files"
+    else:
+        assert wandb_run_dir.name == "files"
+        files_dir = wandb_run_dir.parent / "local-files"
+    files_dir.mkdir(exist_ok=True)
     return files_dir
 
 
